Fragment apps file into command chunks
